+++ src/components/leaderboard/CreateLeaderboardForm.vue
@@ -120,7 +120,7 @@ async function handleSubmit() {
     successMessage.value = `${createdBoard.title} has been created.`;
     await wait(1 * 1000);
 
-    emit('formSuccess');
+    emit('formSuccess', { board: createdBoard });
   } catch {
     errorMessage.value = 'Could not create the leaderboard: something went wrong server-side.';