+++ src/pages/leaderboards/NewLeaderboardPage.vue
@@ -2,17 +2,25 @@
 import { useRouter } from 'vue-router';
 const router = useRouter();
 
+import type { Leaderboard } from 'server/api/v1/leaderboard';
+
 import ApplicationLayout from 'src/layouts/ApplicationLayout.vue';
-import type { MenuItem } from 'primevue/menuitem';
-
 import CreateLeaderboardForm from 'src/components/leaderboard/CreateLeaderboardForm.vue';
-
+import type { MenuItem } from 'primevue/menuitem';
 import Panel from 'primevue/panel';
 
 const breadcrumbs: MenuItem[] = [
   { label: 'Leaderboards', url: '/leaderboards' },
   { label: 'New', url: '/leaderboards/new' },
 ];
+
+function handleFormSuccess({ board }: { board: Leaderboard }) {
+  router.push({ name: 'leaderboard', params: { boardUuid: board.uuid } });
+}
+
+function handleFormFailure() {
+  router.push({ name: 'leaderboards' });
+}
 </script>
 
 <template>
@@ -23,8 +31,8 @@ const breadcrumbs: MenuItem[] = [
       header="Create a new Leaderboard"
     >
       <CreateLeaderboardForm
-        @form-success="router.push({ name: 'leaderboards' })"
-        @form-cancel="router.push({ name: 'leaderboards' })"
+        @form-success="handleFormSuccess"
+        @form-cancel="handleFormFailure"
       />
     </Panel>
   </ApplicationLayout>